# Optimize `new_data_mask()` and `r_alloc_environment()`