--- a/src/rlang/decl/env-decl.h
+++ b/src/rlang/decl/env-decl.h
@@ -2,6 +2,7 @@ r_obj* eval_with_x(r_obj* call, r_obj* x);
 r_obj* eval_with_xy(r_obj* call, r_obj* x, r_obj* y);
 r_obj* eval_with_xyz(r_obj* call, r_obj* x, r_obj* y, r_obj* z);
 
+#if R_VERSION < R_Version(4, 1, 0)
 static
 r_obj* new_env_call;
 
@@ -10,6 +11,7 @@ r_obj* new_env__parent_node;
 
 static
 r_obj* new_env__size_node;
+#endif
 
 static
 r_obj* exists_call;